feat(arcane/physics): mouse-joint maxForce -> 1e6 (MKS) + joints tests -> MKS (MKS P5)
ENGINE -- the MouseJoint maxForce trio (the last 3 MKS-DEFER markers in the repo):
1e9 -> 1e6 at all three sites, MKS-DEFER(P5) markers deleted. MKS-DEFER count is
now ZERO tree-wide.
  * Joint.hpp:100   JointDef struct default (the only reachable site) -- DEFER
                    marker replaced with a real doc comment.
  * Joints.hpp:166  MouseJoint::m_maxForce dead default (ctor always sets it).
  * Joints.cpp:458  "0 -> engine default" ternary fallback.
1e6 N = Box2D's own drag-sample convention 1000*mass*g (samples/sample.cpp:338)
at the heaviest in-range body (~100 kg, g=10): Box2D-derived AND "effectively
unclamped" for all in-range content. DEAD in production -- the only MouseJoint
construction (PhysicsJointsTest) overrides maxForce explicitly, and Sandbox drag
is a separate hand-rolled bounded-impulse mechanism, not a MouseJoint. Zero blast
radius. Documented alternative (Box2D's b2DefaultMouseJointDef().maxForce = 1.0f,
a deliberately-unusable placeholder) rejected here as higher forward-risk for zero
current benefit; recorded for merge-review override.

TEST -- PhysicsJointsTest.cpp converted to MKS (22 PX-PINs -> 0). Geometry /10;
Rule-3 force/torque split (all bodies density=1, /10 length):
  linear force/impulse /1000 (mass /100 x velocity/accel /10):
    prismatic impulse 2000 -> named mass*dv (0.785 kg * 2.5 m/s ~= 1.96 N*s)
    mouse maxForce    8000 -> 8 N
  torque /10000 (inertia /100*/100; angular-accel rad/s^2 length-independent):
    motor  maxMotorTorque 1e6 -> 100 N*m
    wheel  maxMotorTorque 1e6 -> 100 N*m
    clamp  maxMotorTorque 1   -> 0.0001 N*m
maxMotorTorque is TEST content (/10000), NOT part of the maxForce trio (its
JointDef default is 0 = motor disabled, no sentinel).

GravityWorld gravityY fix (survey ERRATA 1): the unmarked wd.gravityY = Real(400)
carried no PX-PIN so a pin-grep would miss it; deleted so gravity inherits the MKS
default +10 (same +Y-down direction) -- affects 7 of 10 cases. Deliberate zero-g
gravityX/gravityY lines (mouse case 3, motor cases 6/7) strip-and-kept.

Length tolerances /10 (3->0.3, 2->0.2, 4->0.4, 8->0.8, >45->>4.5, >28->>2.8,
<120->12, wheel travel >0.5->>0.05); radian tolerances (<0.1, >0.5 angle) and
rad/s bounds (margin 0.5, rate < 2) kept verbatim.

Motor-rate bit-identity CONFIRMED (survey 1d): angularAccel = torque/I with BOTH
/10000 leaves the measured rate numerically unchanged, so cases 6/7's rad/s bounds
stayed bit-identical -- verified, held on the first run, no re-baseline.

Settle counts (900/600/240/600/300/120/600) re-verified empirically -- all 636
[joints] assertions cleared on the first run at MKS geometry + g=10 with the
original loop counts; no count or tolerance re-baselined.

Build: 0W/0E. [joints] 636 assertions / 10 cases green. [physics] 30638 / 278
green (0/0 delta vs post-Task-2 baseline).

+++ Core/src/Arcane/Physics/Joints/Joint.hpp
@@ -95,9 +95,12 @@ namespace Arcane
             // Prismatic / Wheel: world axis (direction). Normalized at Prepare.
             Vec2 axis{ Real(1), Real(0) };
 
-            // Mouse: target + force clamp.
+            // Mouse: target + force clamp. Default is an MKS-honest "effectively unclamped"
+            // value = Box2D's drag-sample convention 1000*mass*g (samples/sample.cpp:338)
+            // at the heaviest in-range body (~100 kg, g=10). Real callers set this per-body;
+            // the only in-repo MouseJoint (PhysicsJointsTest) overrides it explicitly.
             Vec2 target{ Real(0), Real(0) };
-            Real maxForce = Real(1e9); // MKS-DEFER(P5): rescale -- at MKS mass*g is ~1-1e3 N, 1e9 is a px-era "infinite" clamp
+            Real maxForce = Real(1e6);
 
             // Wheel suspension spring (b2WheelJoint). frequencyHz <= 0 -> a rigid
             // axis constraint (no suspension travel). dampingRatio is the spring's
